Reorganize project
- Updated username in all instances
- Renamed top-level package to be more appropriate
- Removed clutter from `mods.toml`
- Updated version number

Renamed classes:
`RegistryHandler` -> `ModItems`
`ModToolTier` -> `ModTiers`
`ModArmorTier` -> `ModArmorMaterials`

All renamed classes have been moved to a new package: `common.items`
