+++ build.gradle
@@ -16,8 +16,8 @@ plugins {
 apply plugin: 'net.minecraftforge.gradle'
 
 
-version = '1.18.2-1.0.0.0'
-group = 'com.shadow_gry.emeraldtools' // http://maven.apache.org/guides/mini/guide-naming-conventions.html
+version = '1.18.2-1.1.0.0'
+group = 'io.github.shadowgry.emeraldtools' // http://maven.apache.org/guides/mini/guide-naming-conventions.html
 archivesBaseName = 'emeraldtools'
 
 // Mojang ships Java 17 to end users in 1.18+, so your mod should target Java 17.
@@ -171,11 +171,11 @@ jar {
     manifest {
         attributes([
                 "Specification-Title"     : "emeraldtools",
-                "Specification-Vendor"    : "emeraldtoolssareus",
+                "Specification-Vendor"    : "ShadowGry",
                 "Specification-Version"   : "1", // We are version 1 of ourselves
                 "Implementation-Title"    : project.name,
                 "Implementation-Version"  : project.jar.archiveVersion,
-                "Implementation-Vendor"   : "emeraldtoolssareus",
+                "Implementation-Vendor"   : "ShadowGry",
                 "Implementation-Timestamp": new Date().format("yyyy-MM-dd'T'HH:mm:ssZ")
         ])
     }
