+++ src/main/java/com/github/shadowgry/emeraldtools/common/items/ModTiers.java
@@ -1,0 +1,29 @@
+/*
+ * This file is part of Emerald Tools.
+ * Copyright (C) 2022  ShadowGry
+ *
+ * Emerald Tools is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+package com.github.shadowgry.emeraldtools.common.items;
+
+import net.minecraft.world.item.Items;
+import net.minecraft.world.item.crafting.Ingredient;
+import net.minecraftforge.common.ForgeTier;
+import net.minecraftforge.common.Tags;
+
+public class ModTiers {
+    public static final ForgeTier EMERALD = new ForgeTier(3, 779, 7.0F, 2.5F, 12, Tags.Blocks.NEEDS_GOLD_TOOL, () -> {
+        return Ingredient.of(Items.EMERALD);
+    });
+}
